+++ CHANGELOG.md
@@ -22,6 +22,8 @@ If MD5 sum is not listed for a certain release then it means that the container 
 ### Added
 
 * Add Conda environment file for project dependencies
+* Add Python packages `scikit-survival, pandas-plink, numba, xmltodict, pyliftover, configparser, intervaltree` to `python3.sif` container
+* Add `Haplin` R package to `r.sif` container
 * Add container build and push actions for all containers:
   * Action should trigger builds on pushes and pull requests targeting the main branch.
   * Should build and push Docker and Singularity images for new tags with `v*.*.*` pattern in main branch.